Helper methods for CDF Physical to Logical Transformation (#579)
## What changes are proposed in this pull request?
This PR introduces methods methods that will be useful to read CDF data
and transform it from its physical form into the logical schema. The
methods are:
- `get_cdf_columns`: Generates a map from cdf column name to expression
that fills that column
- `physical_to_logical_expression`: Generates the physical to logical
expression used to transform the engine data
- `scan_file_read_schema`: Gets the physical schema. This depends on the
cdf scan file type

We also introduce the method `Scalar::timestamp_ntz_from_millis` which
converts from an `i64` millisecond value to a `Scalar::TimestampNtz`.

## How was this change tested?
We test that `physical_to_logical_expression` generates the correct
expression for a `CdfScanFile` with:
- partition columns
- normal selected columns
- generate expression for the `_change_type` column in the case of add,
cdc, and remove CdfScanFile.