Add fix for sv extension (#591)
## What changes are proposed in this pull request?
This is a PR fixes the behaviour for CDF on tables with deletion
vectors.

When computing selection vectors from add/remove dv pairs, the selection
vector must be extended with false.

When computing selection vectors from a normal dv, we extend the
selection vector with true.
## How was this change tested?
I create a table which adds and removes deletion vectors whose lengths
are shorter than the total number of rows. The test checks:
- deleting rows when none were deleted. 
- deleting rows when some were deleted
- restoring some of the rows that have been deleted
- restoring some rows and removing another
- restoring all rows.
All of these rows are less than the size of the table, so they exercise
the extension behaviour for selection vectors.