[fix] move `parquet_stats_skipping` to `predicates` module (#602)
Our builds are actually broken due to a dependency on `mod engine` in
`scan/mod.rs` (which only exists when one of the engine feature-flags is
enabled. as soon as all are off, it fails to build). This fixes the
issue by moving the `parquet_stats_skipping` module out of engine code
and into `expressions` module. This was trivial since it actually
doesn't depend on any engine code.

We pull `parquet_stats_skipping` into `scan/mod.rs` in order to `impl
ParquetStatsProvider for NoStats`. Now instead of having this rely on
engine code, it's moved to the `predicates` module (which is always
included)

Note that I have a new test in CI that will catch this failure in the
future: #601